requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
undetected-chromedriver==3.5.5
selenium==4.25.0
beautifulsoup4==4.12.3
requests==2.32.3
curl-cffi==0.7.1
google-generativeai==0.8.3
Pillow==10.4.0
apscheduler==3.10.4
aiofiles==24.1.0
